+++ testlib/src/main/kotlin/ru/testing/testlib/domain/User.kt
@@ -1,9 +1,7 @@
 package ru.testing.testlib.domain
 
-// TODO: We need a domain module, probably.
-// Placing it into polygon will result in cyclic dependency between configuration and polygon.
 data class User(
-    val id: Int,
+    val id: Long,
     val name: String,
     val passwordHash: String
 )
